+++ packages/flow-table/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./components/f-table/f-table";
+export * from "./components/f-trow/f-trow";
+export * from "./components/f-tcell/f-tcell";
+export * from "./components/f-table-schema/f-table-schema";